feat: store query details
