+++ package/src/swagger_server/controllers/helpers.py
@@ -10,3 +10,10 @@ def check_id(f):
             return {"ERROR": "Invalid id"}, 400
 
     return wrapper
+
+def _stringify_object_id(result):
+    stringified_result = []
+    for element in result:
+        element['_id'] = str(element['_id'])
+        stringified_result.append(element)
+    return stringified_result
