+++ package/src/swagger_server/swagger/swagger.yaml
@@ -309,6 +309,25 @@ paths:
               $ref: '#/components/schemas/Parameter'
         description: name of data field and new value
         required: true
+  /query/post:
+    post:
+      tags:
+        - query
+      summary: store query details
+      description: ''
+      operationId: swagger_server.controllers.customer_db_functions.post_query
+      responses:
+        '200':
+          description: OK
+        '401':
+          description: Authentication error
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: '#/components/schemas/Query'
+        description: query object
+        required: true
 
 servers:
   - url: 'http://localhost:8000/v0.1'
@@ -380,17 +399,11 @@ components:
         payment:
           type: object
           properties:
-            price:
-              type: object
-              properties:
-                amount:
-                  type: string
-                currency:
-                  type: string
+            currency:
+              type: string
             complete:
-              type: string
-            timestamp:
-              type: string
+              type: boolean
+              default: false
       xml:
         name: Payment
     Parameter:
@@ -402,3 +415,61 @@ components:
           type: string
       xml:
         name: Parameter
+    Query:
+      type: object
+      properties:
+        id:
+          type: string
+        customer_id:
+          type: string
+        timestamp:
+          type: string
+        event:
+          type: object
+          properties:
+            type:
+              type: string
+            name:
+              type: string
+            address:
+              type: string
+            start_date:
+              type: string
+            end_date:
+              type: string
+            info:
+              type: string
+            flyer:
+              type: string
+            message:
+              type: string
+        query:
+          type: object
+          properties:
+            details:
+              type: array
+              items:
+                type: object
+                properties:
+                  university_id: 
+                    type: string
+                  university_name: 
+                    type: string
+                  faculty_id: 
+                    type: string
+                  faculty_name: 
+                    type: string
+                  degree_id:
+                    type: string
+                  degree_name:
+                    type: string
+                  course_id:
+                    type: string
+                  course_name:
+                    type: string
+                  absolute:
+                    type: integer
+                  percentage:
+                    type: integer
+      xml:
+        name: Query
